refactor: centralize file path config
