+++ ModuleFolders/Domain/FileReader/ReaderUtil.py
@@ -11,7 +11,7 @@ from bs4 import BeautifulSoup
 from mediapipe.tasks.python import text, BaseOptions
 from mediapipe.tasks.python.text import LanguageDetector
 
-from ModuleFolders.Infrastructure.Platform.PlatformPaths import resource_path
+from ModuleFolders.Config.FilePathConfig import resource_path
 from ModuleFolders.Service.Cache.CacheFile import CacheFile
 from ModuleFolders.Service.Cache.CacheItem import CacheItem
