+++ ModuleFolders/Domain/PromptBuilder/PromptBuilder.py
@@ -3,7 +3,7 @@ from types import SimpleNamespace
 
 from ModuleFolders.Base.Base import Base
 from ModuleFolders.Service.TaskExecutor import TranslatorUtil
-from ModuleFolders.Infrastructure.Platform.PlatformPaths import prompt_path
+from ModuleFolders.Config.FilePathConfig import prompt_path
 from ModuleFolders.Infrastructure.TaskConfig.TaskConfig import TaskConfig
 from ModuleFolders.Domain.PromptBuilder.PromptBuilderEnum import PromptBuilderEnum
 class PromptBuilder(Base):
