+++ ModuleFolders/Domain/PromptBuilder/PromptBuilderLocal.py
@@ -4,7 +4,7 @@ from types import SimpleNamespace
 from ModuleFolders.Base.Base import Base
 from ModuleFolders.Domain.PromptBuilder.PromptBuilder import PromptBuilder
 from ModuleFolders.Domain.PromptBuilder.PromptBuilderEnum import PromptBuilderEnum
-from ModuleFolders.Infrastructure.Platform.PlatformPaths import prompt_path
+from ModuleFolders.Config.FilePathConfig import prompt_path
 from ModuleFolders.Infrastructure.TaskConfig.TaskConfig import TaskConfig
 from ModuleFolders.Service.TaskExecutor import TranslatorUtil
