+++ ModuleFolders/Domain/PromptBuilder/PromptBuilderPolishing.py
@@ -3,7 +3,7 @@ import re
 from ModuleFolders.Base.Base import Base
 from ModuleFolders.Domain.PromptBuilder.PromptBuilder import PromptBuilder
 from ModuleFolders.Domain.PromptBuilder.PromptBuilderEnum import PromptBuilderEnum
-from ModuleFolders.Infrastructure.Platform.PlatformPaths import prompt_path
+from ModuleFolders.Config.FilePathConfig import prompt_path
 from ModuleFolders.Infrastructure.TaskConfig.TaskConfig import TaskConfig
