+++ ModuleFolders/Infrastructure/Platform/__init__.py
@@ -1,1 +1,1 @@
-"""Platform-specific helpers for desktop packaging and runtime paths."""
+"""Platform-specific helpers for desktop packaging and runtime setup."""
