+++ UserInterface/Platform/PlatformPage.py
@@ -24,7 +24,7 @@ from qfluentwidgets import (
 
 from ModuleFolders.Base.Base import Base
 from ModuleFolders.Config.Config import ConfigMixin
-from ModuleFolders.Infrastructure.Platform.PlatformPaths import platform_preset_path
+from ModuleFolders.Config.FilePathConfig import platform_preset_path
 from UserInterface.Platform.APIBindingDialog import APIBindingDialog
 from UserInterface.Platform.APIEditPage import APIEditPage
 from UserInterface.Platform.APIItemCard import APIItemCard
